Move `evaluate_constraint_factor` to `Formulation` utils
